Check productlisting with opensearch tenant
