--- src/FilterService/FilterType/ElasticSearch/Select.php
@@ -1,53 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * Pimcore
- *
- * This source file is available under two different licenses:
- * - GNU General Public License version 3 (GPLv3)
- * - Pimcore Commercial License (PCL)
- * Full copyright and license information is available in
- * LICENSE.md which is distributed with this source code.
- *
- *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- *  @license    http://www.pimcore.org/license     GPLv3 and PCL
- */
-
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\ElasticSearch;
-
-use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
-use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
-use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-
-class Select extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\Select
-{
-    public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
-    {
-        $productList->prepareGroupByValues($this->getField($filterDefinition), true);
-    }
-
-    public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
-    {
-        $field = $this->getField($filterDefinition);
-        $preSelect = $this->getPreSelect($filterDefinition);
-
-        $value = $params[$field] ?? null;
-        $isReload = $params['is_reload'] ?? null;
-
-        if ($value == AbstractFilterType::EMPTY_STRING) {
-            $value = null;
-        } elseif (empty($value) && !$isReload) {
-            $value = $preSelect;
-        }
-
-        $value = trim((string)$value);
-        $currentFilter[$field] = $value;
-
-        if (!empty($value)) {
-            $productList->addCondition(trim($value), $field);
-        }
-
-        return $currentFilter;
-    }
-}
