--- src/FilterService/FilterType/ElasticSearch/SelectCategory.php
@@ -1,82 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * Pimcore
- *
- * This source file is available under two different licenses:
- * - GNU General Public License version 3 (GPLv3)
- * - Pimcore Commercial License (PCL)
- * Full copyright and license information is available in
- * LICENSE.md which is distributed with this source code.
- *
- *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- *  @license    http://www.pimcore.org/license     GPLv3 and PCL
- */
-
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\ElasticSearch;
-
-use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
-use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
-use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-use Pimcore\Model\DataObject\Fieldcollection\Data\FilterCategory;
-use Pimcore\Model\Element\ElementInterface;
-
-class SelectCategory extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SelectCategory
-{
-    public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
-    {
-        $productList->prepareGroupBySystemValues($filterDefinition->getField(), true);
-    }
-
-    /**
-     * @param FilterCategory $filterDefinition
-     *
-     * @throws \Exception
-     */
-    public function getFilterValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter): array
-    {
-        $rawValues = $productList->getGroupBySystemValues($filterDefinition->getField(), true);
-        $values = [];
-
-        foreach ($rawValues as $v) {
-            $values[$v['value']] = ['value' => $v['value'], 'count' => $v['count']];
-        }
-
-        return [
-            'hideFilter' => $filterDefinition->getRequiredFilterField() && empty($currentFilter[$filterDefinition->getRequiredFilterField()]),
-            'label' => $filterDefinition->getLabel(),
-            'currentValue' => $currentFilter[$filterDefinition->getField()],
-            'values' => array_values($values),
-            'indexedValues' => $values,
-            'document' => $this->request->get('contentDocument'),
-            'fieldname' => $filterDefinition->getField(),
-            'rootCategory' => method_exists($filterDefinition, 'getRootCategory') ? $filterDefinition->getRootCategory() : null,
-            'resultCount' => $productList->count(),
-        ];
-    }
-
-    public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
-    {
-        $value = $params[$filterDefinition->getField()] ?? null;
-        $isReload = $params['is_reload'] ?? null;
-
-        if ($value == AbstractFilterType::EMPTY_STRING) {
-            $value = null;
-        } elseif (empty($value) && !$isReload && method_exists($filterDefinition, 'getPreSelect')) {
-            $value = $filterDefinition->getPreSelect();
-            if ($value instanceof ElementInterface) {
-                $value = (string)$value->getId();
-            }
-        }
-
-        $currentFilter[$filterDefinition->getField()] = $value;
-
-        if (!empty($value)) {
-            $value = trim($value);
-            $productList->addCondition($value, $filterDefinition->getField());
-        }
-
-        return $currentFilter;
-    }
-}
