+++ src/FilterService/FilterType/SearchIndex/MultiSelect.php
@@ -14,33 +14,30 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\OpenSearch;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SearchIndex;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\Exception\InvalidConfigException;
 use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
+use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ElasticSearch\AbstractElasticSearch;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-use Pimcore\Model\DataObject\Fieldcollection\Data\FilterMultiSelectFromMultiSelect;
+use Pimcore\Model\DataObject\Fieldcollection\Data\FilterMultiSelect;
 
-class MultiSelectFromMultiSelect extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\MultiSelectFromMultiSelect
+class MultiSelect extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\MultiSelect
 {
     public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
     {
-        $useAndCondition = false;
-
-        if (!$filterDefinition instanceof FilterMultiSelectFromMultiSelect) {
+        if (!$filterDefinition instanceof FilterMultiSelect) {
             throw new InvalidConfigException('invalid configuration');
         }
 
-        $useAndCondition = $filterDefinition->getUseAndCondition();
-
         $field = $this->getField($filterDefinition);
-        $productList->prepareGroupByValues($field, true, !$useAndCondition);
+        $productList->prepareGroupByValues($field, true, !$filterDefinition->getUseAndCondition());
     }
 
     /**
+     * @param FilterMultiSelect $filterDefinition
      *
-     * @return string[]
      */
     public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
     {
@@ -50,17 +47,15 @@ class MultiSelectFromMultiSelect extends \Pimcore\Bundle\EcommerceFrameworkBundl
         $value = $params[$field] ?? null;
         $isReload = $params['is_reload'] ?? null;
 
-        if (empty($value) && !$isReload) {
-            if (is_array($preSelect)) {
-                $value = $preSelect;
-            } else {
-                $value = explode(',', $preSelect);
+        if (!empty($value)) {
+            if (!is_array($value)) {
+                $value = [$value];
             }
+        }
 
-            foreach ($value as $key => $v) {
-                if (!$v) {
-                    unset($value[$key]);
-                }
+        if (empty($value) && !$isReload) {
+            if (!empty($preSelect) || $preSelect == '0') {
+                $value = explode(',', $preSelect);
             }
         } elseif (!empty($value) && in_array(AbstractFilterType::EMPTY_STRING, $value)) {
             $value = null;
@@ -69,21 +64,33 @@ class MultiSelectFromMultiSelect extends \Pimcore\Bundle\EcommerceFrameworkBundl
         $currentFilter[$field] = $value;
 
         if (!empty($value)) {
-            if (!$filterDefinition instanceof FilterMultiSelectFromMultiSelect) {
+            $quotedValues = [];
+            foreach ($value as $v) {
+                if (!empty($v)) {
+                    $quotedValues[] = $v;
+                }
+            }
+
+            if (!$productList instanceof AbstractElasticSearch) {
                 throw new InvalidConfigException('invalid configuration');
             }
 
-            if ($filterDefinition->getUseAndCondition()) {
-                foreach ($value as $entry) {
-                    $productList->addCondition(['term' => ['attributes.' . $field => $entry]], $field);
+            $tenantConfig = $productList->getTenantConfig();
+            $attributeConfig = $tenantConfig->getAttributeConfig()[$field];
+            if ($attributeConfig['type'] == 'boolean') {
+                foreach ($quotedValues as $k => $v) {
+                    $quotedValues[$k] = (bool)$v;
                 }
-            } else {
-                $boolArray = [];
-                foreach ($value as $entry) {
-                    $boolArray[] = ['term' => ['attributes.' . $field => $entry]];
-                }
+            }
 
-                $productList->addCondition(['bool' => ['should' => $boolArray, 'minimum_should_match' => 1]], $field);
+            if (!empty($quotedValues)) {
+                if ($filterDefinition->getUseAndCondition()) {
+                    foreach ($quotedValues as $value) {
+                        $productList->addCondition($value, $field);
+                    }
+                } else {
+                    $productList->addCondition(['terms' => ['attributes.' . $field => $quotedValues]], $field);
+                }
             }
         }
