+++ src/FilterService/FilterType/SearchIndex/SelectFromMultiSelect.php
@@ -14,51 +14,41 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\OpenSearch;
+namespace Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SearchIndex;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\AbstractFilterType;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\ProductList\ProductListInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Model\AbstractFilterDefinitionType;
-use Pimcore\Model\DataObject\Fieldcollection\Data\FilterNumberRangeSelection;
 
-class NumberRangeSelection extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\NumberRangeSelection
+class SelectFromMultiSelect extends \Pimcore\Bundle\EcommerceFrameworkBundle\FilterService\FilterType\SelectFromMultiSelect
 {
     public function prepareGroupByValues(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList): void
     {
         $productList->prepareGroupByValues($this->getField($filterDefinition), true);
     }
 
-    /**
-     * @param FilterNumberRangeSelection $filterDefinition
-     *
-     */
     public function addCondition(AbstractFilterDefinitionType $filterDefinition, ProductListInterface $productList, array $currentFilter, array $params, bool $isPrecondition = false): array
     {
         $field = $this->getField($filterDefinition);
-        $rawValue = $params[$field] ?? null;
+        $preSelect = $this->getPreSelect($filterDefinition);
 
-        if (!empty($rawValue) && $rawValue != AbstractFilterType::EMPTY_STRING) {
-            $values = explode('-', $rawValue);
-            $value['from'] = trim($values[0]);
-            $value['to'] = trim($values[1]);
-        } elseif ($rawValue == AbstractFilterType::EMPTY_STRING) {
+        $value = $params[$field] ?? null;
+        $isReload = $params['is_reload'] ?? null;
+
+        if ($value == AbstractFilterType::EMPTY_STRING) {
             $value = null;
-        } else {
-            $value['from'] = $filterDefinition->getPreSelectFrom();
-            $value['to'] = $filterDefinition->getPreSelectTo();
+        } elseif (empty($value) && !$isReload) {
+            $value = $preSelect;
+        }
+
+        if (!empty($value)) {
+            $value = trim($value);
         }
 
         $currentFilter[$field] = $value;
 
-        if (!empty($value) && ($value['from'] !== null || $value['to'] !== null)) {
-            $range = [];
-            if (!empty($value['from'])) {
-                $range['gte'] = $value['from'];
-            }
-            if (!empty($value['to'])) {
-                $range['lte'] = $value['to'];
-            }
-            $productList->addCondition(['range' => ['attributes.' . $field => $range]], $field);
+        if (!empty($value)) {
+            $productList->addCondition(['term' => ['attributes.' . $field => $value]], $field);
         }
 
         return $currentFilter;
